+++ Changelog.md
@@ -61,6 +61,7 @@
 * Add parsing 'DocxParagraph#inserted' context
 * Add parsing `bgPr` and `stretch`
 * Add parsing `DisplayLabelsProperties` in `Series`
+* Add parsing `FootnoteProperties`
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit